add changed to uint16

+++ relay-node/list_utility.h
@@ -6,15 +6,15 @@
 
 typedef struct child_list_struct {
   struct child_list_struct *next;
-  uint8_t child_id;
+  uint16_t child_id;
   uint8_t received_pkt;
 }child_node_t;
 
 void init_child_list(void);
-uint8_t is_child_set(uint8_t child_id);
+uint8_t is_child_set(uint16_t child_id);
 void rest_child_list();
 uint8_t is_all_child_set();
-uint8_t is_child(uint8_t child_id);
+uint8_t is_child(uint16_t child_id);
 uint8_t get_child_count();
-uint8_t child_set_pkt_received(uint8_t child_id);
+uint8_t child_set_pkt_received(uint16_t child_id);
 void print_child_list(void);
